Split operator poll method into more granular ones
This commit replaces the CameraRigMixin class that only implemented a
poll method, with two poll methods, a base one which behaves the same
as before, and a perspective one which disallows the operator with 2D
rigs and ortho cameras.
The poll methods are now explicitly defined in each operator and UI
class.