2D rig: Use circle for offset control, make widgets more symmetrical
The offset widget is not appropriate for the 2D rig, so use a circle
with another radius instead.

In addition, the 2D rig now uses some of the same widgets as the other
two, but its bones are rotated 90°. This revealed that some of the
widgets were not completely symmetrical.