2D rig: avoid driver error if both corners at same location
The camera drivers for focal length and shift use the distance between
corners as divisor. If they are at the same location, the distance is
zero and a division by zero error will occur. Default to 1.0 in that
case, which doesn't do much but won't make the driver permanently
invalid.